Fetch tags with --force due to changes in Git 2.20
"Until Git version 2.20, and unlike when pushing with git-push[1], any
updates to refs/tags/* would be accepted without + in the refspec (or
--force). When fetching, we promiscuously considered all tag updates
from a remote to be forced fetches. Since Git version 2.20, fetching
to update refs/tags/* works the same way as when pushing. I.e. any
updates will be rejected without + in the refspec (or --force)."

https://git-scm.com/docs/git-fetch
https://github.com/git/git/blob/master/Documentation/RelNotes/2.20.0.txt#L67-L71